+++ kipper/core/src/compiler/ast/nodes/expressions/member-access-expression/member-access-expression.ts
@@ -16,6 +16,7 @@ import type { CompilableASTNode } from "../../../compilable-ast-node";
 import { Expression } from "../expression";
 import { KipperNotImplementedError, UnableToDetermineSemanticDataError } from "../../../../../errors";
 import { kipperInternalBuiltInFunctions } from "../../../../semantics";
+import { ClassMethodDeclaration, ClassPropertyDeclaration } from "../../declarations";
 
 /**
  * A union of all possible {@link KipperParserRuleContext} rule contexts that {@link MemberAccessExpression} implements.
@@ -101,11 +102,18 @@ export class MemberAccessExpression extends Expression<
 	public async primarySemanticAnalysis(): Promise<void> {
 		// Handle the different types of member access expressions
 		if (this.antlrRuleCtx instanceof DotNotationMemberAccessExpressionContext) {
-			throw this.programCtx
-				.semanticCheck(this)
-				.notImplementedError(
-					new KipperNotImplementedError("Member access expressions using dot notation are not yet implemented"),
-				);
+			const antlrChildren = this.getAntlrRuleChildren();
+			const objExp = this.children[0];
+			if (!antlrChildren || !objExp) {
+				throw new UnableToDetermineSemanticDataError();
+			}
+
+			const identifier = antlrChildren.pop()!.text.slice(1);
+			this.semanticData = {
+				objectLike: objExp,
+				propertyIndexOrKeyOrSlice: identifier,
+				accessType: "dot",
+			};
 		} else if (this.antlrRuleCtx instanceof BracketNotationMemberAccessExpressionContext) {
 			const objExp = this.children[0];
 			const keyExp = this.children[1];