+++ kipper/core/src/compiler/semantics/types/built-in/func.ts
@@ -4,7 +4,6 @@ import {
 	ArgumentAssignmentTypeError,
 	AssignmentTypeError,
 	GenericArgumentTypeError,
-	KipperInternalError,
 	MismatchingArgCountBetweenFuncTypesError,
 	PropertyAssignmentTypeError,
 	type TypeError,